move scoping into separate module
