+++ relay-cardinality/src/redis/limiter.rs
@@ -7,23 +7,13 @@ use crate::{
     limiter::{Entry, EntryId, Limiter, Rejections, Scoping},
     redis::{
         cache::{Cache, CacheOutcome},
+        quota::QuotaScoping,
         script::{CardinalityScript, Status},
     },
     statsd::{CardinalityLimiterCounters, CardinalityLimiterHistograms, CardinalityLimiterTimers},
-    window::Slot,
-    CardinalityLimit, CardinalityScope, OrganizationId, Result, SlidingWindow,
+    CardinalityLimit, Result,
 };
-use relay_base_schema::metrics::MetricNamespace;
-use relay_base_schema::project::ProjectId;
 use relay_common::time::UnixTimestamp;
-
-/// Key prefix used for Redis keys.
-const KEY_PREFIX: &str = "relay:cardinality";
-/// Redis key version.
-///
-/// The version is embedded in the key as a static segment, increment the version whenever there are
-/// breaking changes made to the keys or storage format in Redis.
-const KEY_VERSION: u32 = 1;
 
 /// Configuration options for the [`RedisSetLimiter`].
 pub struct RedisSetLimiterOptions {
@@ -177,77 +167,6 @@ impl Limiter for RedisSetLimiter {
     }
 }
 
-/// A quota scoping extracted from a [`CardinalityLimit`] and a [`Scoping`].
-#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq)]
-pub struct QuotaScoping {
-    pub namespace: Option<MetricNamespace>,
-    pub organization_id: Option<OrganizationId>,
-    pub project_id: Option<ProjectId>,
-    window: SlidingWindow,
-}
-
-impl QuotaScoping {
-    /// Creates a new [`QuotaScoping`] from a [`Scoping`] and [`CardinalityLimit`].
-    ///
-    /// Returns `None` for limits with scope [`CardinalityScope::Unknown`].
-    pub fn new(scoping: Scoping, limit: &CardinalityLimit) -> Option<Self> {
-        let (organization_id, project_id) = match limit.scope {
-            CardinalityScope::Organization => (Some(scoping.organization_id), None),
-            // Invalid/unknown scope -> ignore the limit.
-            CardinalityScope::Unknown => return None,
-        };
-
-        Some(Self {
-            window: limit.window,
-            namespace: limit.namespace,
-            organization_id,
-            project_id,
-        })
-    }
-
-    /// Wether the scoping applies to the passed entry.
-    pub fn matches(&self, entry: &Entry) -> bool {
-        self.namespace.is_none() || self.namespace == Some(entry.namespace)
-    }
-
-    /// Returns the currently active slot.
-    pub fn active_slot(&self, timestamp: UnixTimestamp) -> Slot {
-        self.window.active_slot(self.shifted(timestamp))
-    }
-
-    /// Returns all slots of the sliding window for a specific timestamp.
-    pub fn slots(&self, timestamp: UnixTimestamp) -> impl Iterator<Item = Slot> {
-        self.window.iter(self.shifted(timestamp))
-    }
-
-    /// Applies a timeshift based on the granularity of the sliding window to the passed timestamp.
-    ///
-    /// The shift is used to evenly distribute cache and Redis operations across
-    /// the sliding window's granule.
-    fn shifted(&self, timestamp: UnixTimestamp) -> UnixTimestamp {
-        let shift = self
-            .organization_id
-            .map(|o| o % self.window.granularity_seconds)
-            .unwrap_or(0);
-
-        UnixTimestamp::from_secs(timestamp.as_secs() + shift)
-    }
-
-    /// Returns the minimum TTL for a Redis key created by [`Self::into_redis_key`].
-    fn redis_key_ttl(&self) -> u64 {
-        self.window.window_seconds
-    }
-
-    /// Turns the scoping into a Redis key for the passed slot.
-    fn into_redis_key(self, slot: Slot) -> String {
-        let organization_id = self.organization_id.unwrap_or(0);
-        let project_id = self.project_id.map(|p| p.value()).unwrap_or(0);
-        let namespace = self.namespace.map(|ns| ns.as_str()).unwrap_or("");
-
-        format!("{KEY_PREFIX}:{KEY_VERSION}:scope-{{{organization_id}-{project_id}-{namespace}}}-{slot}")
-    }
-}
-
 /// Internal state combining relevant entries for the respective quota.
 struct LimitState<'a> {
     /// Id of the original limit.
@@ -372,6 +291,7 @@ mod tests {
     use relay_redis::{redis, RedisConfigOptions};
 
     use crate::limiter::EntryId;
+    use crate::redis::{KEY_PREFIX, KEY_VERSION};
     use crate::{CardinalityScope, SlidingWindow};
 
     use super::*;
